feat: modify map_page to three files
